requirements.txt
distro==1.8.0
docx2txt==0.8
install==1.3.5
numpy==1.25.2
pandas==2.1.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabula-py==2.7.0
tzdata==2023.3
